requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Werkzeug==2.3.7
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
Pillow==10.1.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
WTForms==3.1.0
